[#2967] Factor out shared code between PC & NPC sheets. Add item description collapsible sections.
