+++ icons/LICENSE
@@ -64,7 +64,7 @@ The dnd5e system for Foundry Virtual Tabletop includes icon artwork licensed fro
 /svg/statuses/transformed.svg - "Flat paw print icon" by Lorc under CC BY 3.0
 
 ==============================================
-SVG Icons from mikiko.art
+Icons from mikiko.art
 ==============================================
 
 The dnd5e system for Foundry Virtual Tabletop includes icon artwork licensed from mikiko.art under CC BY-NC-SA 4.0. These icons are packaged with and provided for use in the dnd5e system under the terms of their license, which can be found here: https://creativecommons.org/licenses/by-nc-sa/4.0/
