+++ less/dnd5e.less
@@ -18,6 +18,7 @@
 
 /* V2 Styles */
 @import "v2/apps.less";
+@import "v2/actors.less";
 @import "v2/character.less";
 @import "v2/inventory.less";
 @import "v2/chat.less";
