+++ less/v2/apps.less
@@ -246,10 +246,16 @@
       &.green { background-color: var(--dnd5e-color-green); }
       &.maroon { background-color: var(--dnd5e-color-maroon); }
 
-      &.pill-sm {
+      &.transparent, &.pill-sm {
         color: #4e4e4e;
         border: 1px dotted var(--color-border-dark-5);
         background-color: transparent;
+      }
+
+      &.pill-xs {
+        font-size: var(--font-size-10);
+        padding: 2px 4px;
+        gap: 3px;
       }
 
       > .separator { color: var(--color-text-light-6); }
@@ -345,7 +351,7 @@
     flex-direction: column;
     gap: .125rem;
 
-    + .meter-group { margin-top: .5rem; }
+    + .meter-group { margin-block-start: .5rem; }
   }
 
   .meter {
@@ -359,6 +365,7 @@
     position: relative;
     display: flex;
     align-items: center;
+    overflow: hidden;
 
     &.progress, .progress {
       --bar-percentage: 0%;
@@ -368,8 +375,8 @@
       &::before {
         content: "";
         position: absolute;
-        height: 100%;
-        width: calc(var(--bar-percentage) - var(--border-width));
+        block-size: 100%;
+        inline-size: calc(var(--bar-percentage) - var(--border-width));
         box-shadow: 0 0 6px var(--dnd5e-shadow-45);
         clip-path: polygon(0 0, calc(100% + 6px) 0, calc(100% + 6px) 100%, 0 100%);
         border-radius: 4px;
@@ -377,13 +384,13 @@
     }
 
     .label {
-      width: 100%;
+      inline-size: 100%;
       display: flex;
       align-items: center;
       gap: .25rem;
       position: relative;
-      padding-inline-start: .5rem;
-      padding-inline-end: .75rem;
+      padding-inline-start: 8px;
+      padding-inline-end: 12px;
       color: var(--color-text-light-0);
 
       &[hidden] { display: none; }
@@ -591,11 +598,12 @@
     padding: 10px 12px;
 
     .pip {
+      --pip-size: 16px;
       background: var(--dnd5e-color-light-gray);
       border: 2px solid var(--dnd5e-color-gold);
       border-radius: 100%;
-      width: 16px;
-      height: 16px;
+      width: var(--pip-size);
+      height: var(--pip-size);
       display: grid;
       place-content: center;
       font-weight: bold;
