+++ less/v2/dark/apps.less
@@ -43,6 +43,11 @@
         border-color: var(--dnd5e-color-blue-gray-1);
         color: var(--dnd5e-color-blue-white);
       }
+
+      &.transparent, &.pill-sm {
+        color: var(--color-text-dark-primary);
+        border-style: solid;
+      }
     }
 
     .pill-lg {
