+++ less/v2/dark/inventory.less
@@ -14,6 +14,8 @@
         &[data-context-menu] { color: var(--color-text-light-6); }
         &.active { color: var(--dnd5e-color-gold); }
       }
+
+      .item-summary { background: transparent; }
     }
 
     .item-controls .info-control { color: var(--dnd5e-color-gold); }
