+++ less/v2/inventory.less
@@ -213,11 +213,13 @@
   .items-section {
 
     .items-header, .item {
-      display: flex;
-      align-items: stretch;
-
       &[hidden] { display: none; }
       .condensed { font-family: var(--dnd5e-font-roboto-condensed); }
+
+      .item-row {
+        display: flex;
+        align-items: stretch;
+      }
 
       .item-name {
         flex: 1;
@@ -392,10 +394,9 @@
       }
 
       /* Spell Target & Formula */
-      .item-target, .item-formula { width: 80px; }
+      .item-target, .item-formula, .spell-formula { width: 80px; }
 
-      .item-formula {
-        display: none;
+      .item-formula, .spell-formula {
         flex-direction: column;
 
         dnd5e-icon {
@@ -408,6 +409,9 @@
           gap: .1875rem;
         }
       }
+
+      .item-formula { display: flex; }
+      .spell-formula { display: none; }
 
       /* Effect Duration */
       .effect-name {
@@ -426,7 +430,7 @@
 
       /* Item Controls */
       .item-controls {
-        width: 60px;
+        width: 70px;
         align-items: stretch;
         justify-content: center;
         gap: .375rem;
@@ -464,17 +468,38 @@
         align-items: center;
       }
 
-      @container (min-width: 600px) {
-        .item-price, .item-formula, .spell-uses { display: flex; }
+      /* Item Description */
+      .item-description {
+        font-size: var(--font-size-11);
+
+        .item-summary {
+          padding: 1px 8px 8px;
+          background: var(--dnd5e-color-parchment);
+        }
+      }
+
+      @container (min-width: 580px) {
+        .spell-uses { display: flex; }
         .subtitle-uses { display: none; }
       }
 
-      @container (min-width: 650px) {
+      @container (min-width: 600px) {
+        .item-price { display: flex; }
+      }
+
+      @container (min-width: 640px) {
+        .spell-formula { display: flex; }
+      }
+
+      @container (min-width: 660px) {
         .item-school { display: flex; }
       }
     }
 
     .items-header {
+      display: flex;
+      align-items: stretch;
+
       .config-button {
         position: absolute;
         inset: .5rem 1.5rem auto auto;
@@ -491,6 +516,11 @@
     .item {
       border-bottom: var(--dnd5e-border-dotted);
       &:last-child { border: none; }
+    }
+
+    .effect {
+      display: flex;
+      align-items: stretch;
     }
   }
