+++ less/variables.less
@@ -35,6 +35,7 @@
   --dnd5e-color-green: #205b15;
   --dnd5e-color-maroon: #741b2b;
   --dnd5e-color-parchment: #f1ebe8;
+  --dnd5e-color-purple: #c44dff;
   --dnd5e-color-scrollbar: var(--dnd5e-color-gold);
   --dnd5e-color-card: #f8f4f1;
   --dnd5e-color-hp-1: #1b4b29;
@@ -61,6 +62,7 @@
   --dnd5e-color-application-damage: #9c5b47;
   --dnd5e-color-application-healing: #3c7f58;
   --dnd5e-color-application-temp: #007F7F;
+  --dnd5e-background-25: rgb(0 0 0 / 25%);
   --dnd5e-background-10: rgb(0 0 0 / 10%);
   --dnd5e-background-5: rgb(0 0 0 / 5%);
   --dnd5e-background-card: var(--dnd5e-color-card);
