+++ module/applications/actor/character-sheet-2.mjs
@@ -1,15 +1,16 @@
 import CharacterData from "../../data/actor/character.mjs";
 import * as Trait from "../../documents/actor/trait.mjs";
-import { formatNumber, simplifyBonus, staticID } from "../../utils.mjs";
+import { simplifyBonus, staticID } from "../../utils.mjs";
 import ContextMenu5e from "../context-menu.mjs";
 import SheetConfig5e from "../sheet-config.mjs";
-import Tabs5e from "../tabs.mjs";
 import ActorSheet5eCharacter from "./character-sheet.mjs";
+import ActorSheetV2Mixin from "./sheet-v2-mixin.mjs";
 
 /**
  * An Actor sheet for player character type actors.
+ * @mixes ActorSheetV2
  */
-export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
+export default class ActorSheet5eCharacter2 extends ActorSheetV2Mixin(ActorSheet5eCharacter) {
   constructor(object, options={}) {
     const key = `character${object.limited ? ":limited" : ""}`;
     const { width, height } = game.user.getFlag("dnd5e", `sheetPrefs.${key}`) ?? {};
@@ -40,15 +41,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   }
 
   /**
-   * Available sheet modes.
-   * @enum {number}
-   */
-  static MODES = {
-    PLAY: 1,
-    EDIT: 2
-  };
-
-  /**
    * Proficiency class names.
    * @enum {string}
    */
@@ -59,18 +51,7 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
     2: "double"
   };
 
-  /**
-   * @typedef {object} SheetTabDescriptor5e
-   * @property {string} tab     The tab key.
-   * @property {string} label   The tab label's localization key.
-   * @property {string} [icon]  A font-awesome icon.
-   * @property {string} [svg]   An SVG icon.
-   */
-
-  /**
-   * Sheet tabs.
-   * @type {SheetTabDescriptor5e[]}
-   */
+  /** @override */
   static TABS = [
     { tab: "details", label: "DND5E.Details", icon: "fas fa-cog" },
     { tab: "inventory", label: "DND5E.Inventory", svg: "backpack" },
@@ -81,25 +62,11 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   ];
 
   /**
-   * The mode the sheet is currently in.
-   * @type {ActorSheet5eCharacter2.MODES}
-   * @protected
-   */
-  _mode = this.constructor.MODES.PLAY;
-
-  /**
    * Whether the user has manually opened the death save tray.
    * @type {boolean}
    * @protected
    */
   _deathTrayOpen = false;
-
-  /**
-   * The cached concentration information for the character.
-   * @type {{items: Set<Item5e>, effects: Set<ActiveEffect5e>}}
-   * @internal
-   */
-  _concentration;
 
   /* -------------------------------------------- */
 
@@ -107,73 +74,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   get template() {
     if ( !game.user.isGM && this.actor.limited ) return "systems/dnd5e/templates/actors/limited-sheet-2.hbs";
     return "systems/dnd5e/templates/actors/character-sheet-2.hbs";
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritDoc */
-  async _renderOuter() {
-    const html = await super._renderOuter();
-    const header = html[0].querySelector(".window-header");
-
-    // Add edit <-> play slide toggle.
-    if ( this.isEditable ) {
-      const toggle = document.createElement("slide-toggle");
-      toggle.checked = this._mode === this.constructor.MODES.EDIT;
-      toggle.classList.add("mode-slider");
-      toggle.dataset.tooltip = "DND5E.SheetModeEdit";
-      toggle.setAttribute("aria-label", game.i18n.localize("DND5E.SheetModeEdit"));
-      toggle.addEventListener("change", this._onChangeSheetMode.bind(this));
-      toggle.addEventListener("dblclick", event => event.stopPropagation());
-      header.insertAdjacentElement("afterbegin", toggle);
-    }
-
-    // Adjust header buttons.
-    header.querySelectorAll(".header-button").forEach(btn => {
-      const label = btn.querySelector(":scope > i").nextSibling;
-      btn.dataset.tooltip = label.textContent;
-      btn.setAttribute("aria-label", label.textContent);
-      label.remove();
-    });
-
-    const idLink = header.querySelector(".document-id-link");
-    if ( idLink ) {
-      const firstButton = header.querySelector(".header-button");
-      firstButton?.insertAdjacentElement("beforebegin", idLink);
-    }
-
-    if ( !game.user.isGM && this.actor.limited ) {
-      html[0].classList.add("limited");
-      return html;
-    }
-
-    // Render tabs.
-    const nav = document.createElement("nav");
-    nav.classList.add("tabs");
-    nav.dataset.group = "primary";
-    nav.append(...this.constructor.TABS.map(({ tab, label, icon, svg }) => {
-      const item = document.createElement("a");
-      item.classList.add("item", "control");
-      item.dataset.group = "primary";
-      item.dataset.tab = tab;
-      item.dataset.tooltip = label;
-      item.setAttribute("aria-label", label);
-      if ( icon ) item.innerHTML = `<i class="${icon}"></i>`;
-      else if ( svg ) item.innerHTML = `<dnd5e-icon src="systems/dnd5e/icons/svg/${svg}.svg"></dnd5e-icon>`;
-      return item;
-    }));
-    html[0].insertAdjacentElement("afterbegin", nav);
-    this._tabs = this.options.tabs.map(t => {
-      t.callback = this._onChangeTab.bind(this);
-      if ( this._tabs?.[0]?.active !== t.initial ) t.initial = this._tabs?.[0]?.active ?? t.initial;
-      return new Tabs5e(t);
-    });
-
-    // Set theme
-    // TODO: Re-enable this when we support V12 only
-    // setTheme(html[0], this.actor.getFlag("dnd5e", "theme"));
-
-    return html;
   }
 
   /* -------------------------------------------- */
@@ -193,33 +93,13 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
 
   /** @inheritDoc */
   async getData(options) {
-    this._concentration = this.actor.concentration; // Cache concentration so it's not called for every item.
     const context = await super.getData(options);
-    context.editable = this.isEditable && (this._mode === this.constructor.MODES.EDIT);
-    context.cssClass = context.editable ? "editable" : this.isEditable ? "interactable" : "locked";
-    const activeTab = (game.user.isGM || !this.actor.limited) ? this._tabs?.[0]?.active ?? "details" : "biography";
-    context.cssClass += ` tab-${activeTab}`;
-    const sidebarCollapsed = game.user.getFlag("dnd5e", `sheetPrefs.character.tabs.${activeTab}.collapseSidebar`);
-    if ( sidebarCollapsed ) {
-      context.cssClass += " collapsed";
-      context.sidebarCollapsed = true;
-    }
     const { attributes, details, traits } = this.actor.system;
 
     // Class
     context.labels.class = Object.values(this.actor.classes).sort((a, b) => {
       return b.system.levels - a.system.levels;
     }).map(c => `${c.name} ${c.system.levels}`).join(" / ");
-
-    // Portrait
-    const showTokenPortrait = this.actor.getFlag("dnd5e", "showTokenPortrait") === true;
-    const token = this.actor.isToken ? this.actor.token : this.actor.prototypeToken;
-    context.portrait = {
-      token: showTokenPortrait,
-      src: showTokenPortrait ? token.texture.src : this.actor.img,
-      // TODO: Not sure the best way to update the parent texture from this sheet if this is a token actor.
-      path: showTokenPortrait ? this.actor.isToken ? "" : "prototypeToken.texture.src" : "img"
-    };
 
     // Exhaustion
     const max = CONFIG.DND5E.conditionTypes.exhaustion.levels;
@@ -243,36 +123,13 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
       return obj;
     }, { value: 0, label: CONFIG.DND5E.movementTypes.walk });
 
-    // Hit Dice
-    context.hd = attributes.hd;
-
     // Death Saves
-    const plurals = new Intl.PluralRules(game.i18n.lang, { type: "ordinal" });
-    context.death = { open: this._deathTrayOpen };
-    ["success", "failure"].forEach(deathSave => {
-      context.death[deathSave] = [];
-      for ( let i = 1; i < 4; i++ ) {
-        const n = deathSave === "failure" ? i : 4 - i;
-        const i18nKey = `DND5E.DeathSave${deathSave.titleCase()}Label`;
-        const filled = attributes.death[deathSave] >= n;
-        const classes = ["pip"];
-        if ( filled ) classes.push("filled");
-        if ( deathSave === "failure" ) classes.push("failure");
-        context.death[deathSave].push({
-          n, filled,
-          tooltip: i18nKey,
-          label: game.i18n.localize(`${i18nKey}N.${plurals.select(n)}`),
-          classes: classes.join(" ")
-        });
-      }
-    });
+    context.death.open = this._deathTrayOpen;
 
     // Ability Scores
     context.abilityRows = Object.entries(context.abilities).reduce((obj, [k, ability]) => {
       ability.key = k;
       ability.abbr = CONFIG.DND5E.abilities[k]?.abbreviation ?? "";
-      ability.sign = Math.sign(ability.mod) < 0 ? "-" : "+";
-      ability.mod = Math.abs(ability.mod);
       ability.baseValue = context.source.abilities[k]?.value ?? 0;
       if ( obj.bottom.length > 5 ) obj.top.push(ability);
       else obj.bottom.push(ability);
@@ -286,8 +143,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
       ability = context.saves[ability.key] = { ...ability };
       ability.class = this.constructor.PROFICIENCY_CLASSES[context.editable ? ability.baseProf : ability.proficient];
       ability.hover = CONFIG.DND5E.proficiencyLevels[ability.proficient];
-      ability.sign = Math.sign(ability.save) < 0 ? "-" : "+";
-      ability.mod = Math.abs(ability.save);
     }
 
     if ( this.actor.statuses.has(CONFIG.specialStatusEffects.CONCENTRATING) || context.editable ) {
@@ -296,8 +151,7 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
         class: "colspan concentration",
         label: game.i18n.localize("DND5E.Concentration"),
         abbr: game.i18n.localize("DND5E.Concentration"),
-        mod: Math.abs(attributes.concentration.save),
-        sign: attributes.concentration.save < 0 ? "-" : "+"
+        save: attributes.concentration.save
       };
     }
 
@@ -311,8 +165,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
     // Skills & Tools
     for ( const [key, entry] of Object.entries(context.skills).concat(Object.entries(context.tools)) ) {
       entry.class = this.constructor.PROFICIENCY_CLASSES[context.editable ? entry.baseValue : entry.value];
-      entry.sign = Math.sign(entry.total) < 0 ? "-" : "+";
-      entry.mod = Math.abs(entry.total);
       if ( key in CONFIG.DND5E.skills ) entry.reference = CONFIG.DND5E.skills[key].reference;
       else if ( key in CONFIG.DND5E.toolIds ) entry.reference = Trait.getBaseItemUUID(CONFIG.DND5E.toolIds[key]);
     }
@@ -332,15 +184,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
     if ( details.background instanceof dnd5e.documents.Item5e ) context.background = details.background;
 
     // Senses
-    context.senses = Object.entries(CONFIG.DND5E.senses).reduce((obj, [k, label]) => {
-      const value = attributes.senses[k];
-      if ( value ) obj[k] = { label, value };
-      return obj;
-    }, {});
-
-    if ( attributes.senses.special ) attributes.senses.special.split(";").forEach((v, i) => {
-      context.senses[`custom${i + 1}`] = { label: v.trim() };
-    });
     if ( foundry.utils.isEmpty(context.senses) ) delete context.senses;
 
     // Spellcasting
@@ -354,64 +197,15 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
       const ability = this.actor.system.abilities[sc.ability];
       const mod = ability?.mod ?? 0;
       const attackBonus = msak === rsak ? msak : 0;
-      const attack = mod + this.actor.system.attributes.prof + attackBonus;
       const name = item.system.spellcasting.progression === sc.progression ? item.name : item.subclass?.name;
       context.spellcasting.push({
         label: game.i18n.format("DND5E.SpellcastingClass", { class: name }),
-        ability: { sign: Math.sign(mod) < 0 ? "-" : "+", value: Math.abs(mod), ability: sc.ability },
-        attack: { sign: Math.sign(attack) < 0 ? "-" : "+", value: Math.abs(attack) },
+        ability: { mod, ability: sc.ability },
+        attack: mod + this.actor.system.attributes.prof + attackBonus,
         primary: this.actor.system.attributes.spellcasting === sc.ability,
         save: ability?.dc ?? 0
       });
     }
-
-    // Containers
-    for ( const container of context.containers ) {
-      const ctx = context.itemContext[container.id];
-      ctx.capacity = await container.system.computeCapacity();
-    }
-
-    // Effects & Conditions
-    const conditionIds = new Set();
-    context.conditions = Object.entries(CONFIG.DND5E.conditionTypes).reduce((arr, [k, c]) => {
-      if ( c.pseudo ) return arr; // Filter out pseudo-conditions.
-      const { label: name, icon, reference } = c;
-      const id = staticID(`dnd5e${k}`);
-      conditionIds.add(id);
-      const existing = this.actor.effects.get(id);
-      const { disabled, img } = existing ?? {};
-      arr.push({
-        name, reference,
-        id: k,
-        icon: img ?? icon,
-        disabled: existing ? disabled : true
-      });
-      return arr;
-    }, []);
-
-    for ( const category of Object.values(context.effects) ) {
-      category.effects = await category.effects.reduce(async (arr, effect) => {
-        effect.updateDuration();
-        if ( conditionIds.has(effect.id) && !effect.duration.remaining ) return arr;
-        const { id, name, img, disabled, duration } = effect;
-        const toggleable = !this._concentration?.effects.has(effect);
-        let source = await effect.getSource();
-        // If the source is an ActiveEffect from another Actor, note the source as that Actor instead.
-        if ( (source instanceof dnd5e.documents.ActiveEffect5e) && (source.target !== this.object) ) {
-          source = source.target;
-        }
-        arr = await arr;
-        arr.push({
-          id, name, img, disabled, duration, source, toggleable,
-          parentId: effect.target === effect.parent ? null : effect.parent.id,
-          durationParts: duration.remaining ? duration.label.split(", ") : [],
-          hasTooltip: source instanceof dnd5e.documents.Item5e
-        });
-        return arr;
-      }, []);
-    }
-
-    context.effects.suppressed.info = context.effects.suppressed.info[0];
 
     // Characteristics
     context.characteristics = [
@@ -433,69 +227,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   /* -------------------------------------------- */
 
   /** @inheritDoc */
-  _getLabels() {
-    const labels = super._getLabels();
-    labels.damageAndHealing = { ...CONFIG.DND5E.damageTypes, ...CONFIG.DND5E.healingTypes };
-    return labels;
-  }
-
-  /* -------------------------------------------- */
-
-  /** @override */
-  _prepareTraits() {
-    const traits = {};
-    for ( const [trait, config] of Object.entries(CONFIG.DND5E.traits) ) {
-      const key = config.actorKeyPath ?? `system.traits.${trait}`;
-      const data = foundry.utils.deepClone(foundry.utils.getProperty(this.actor, key));
-      if ( !data ) continue;
-      let values = data.value;
-      if ( !values ) values = [];
-      else if ( values instanceof Set ) values = Array.from(values);
-      else if ( !Array.isArray(values) ) values = [values];
-      values = values.map(key => {
-        const value = { label: Trait.keyLabel(key, { trait }) ?? key };
-        const icons = value.icons = [];
-        if ( data.bypasses?.size && CONFIG.DND5E.damageTypes[key]?.isPhysical ) icons.push(...data.bypasses);
-        return value;
-      });
-      if ( data.custom ) data.custom.split(";").forEach(v => values.push({ label: v.trim() }));
-      if ( values.length ) traits[trait] = values;
-    }
-    // If petrified, display "All Damage" instead of all damage types separately
-    if ( this.document.hasConditionEffect("petrification") ) {
-      traits.dr = [{ label: game.i18n.localize("DND5E.DamageAll") }];
-    }
-    // Combine damage & condition immunities in play mode.
-    if ( (this._mode === this.constructor.MODES.PLAY) && traits.ci ) {
-      traits.di ??= [];
-      traits.di.push(...traits.ci);
-      delete traits.ci;
-    }
-
-    // Prepare damage modifications
-    const dm = this.actor.system.traits?.dm;
-    if ( dm ) {
-      const rollData = this.actor.getRollData({ deterministic: true });
-      const values = Object.entries(dm.amount).map(([k, v]) => {
-        const total = simplifyBonus(v, rollData);
-        if ( !total ) return null;
-        const value = {
-          label: `${CONFIG.DND5E.damageTypes[k]?.label ?? k} ${formatNumber(total, { signDisplay: "always" })}`,
-          color: total > 0 ? "maroon" : "green"
-        };
-        const icons = value.icons = [];
-        if ( dm.bypasses.size && CONFIG.DND5E.damageTypes[k]?.isPhysical ) icons.push(...dm.bypasses);
-        return value;
-      }).filter(f => f);
-      if ( values.length ) traits.dm = values;
-    }
-
-    return traits;
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritDoc */
   _prepareItems(context) {
     super._prepareItems(context);
     context.containers = context.inventory
@@ -505,15 +236,15 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
     context.inventory.push({ label: "DND5E.Contents", items: [], dataset: { type: "all" } });
 
     // Remove races & background as they are shown on the details tab instead.
-    context.features = context.features.filter(f => (f.dataset.type !== "background") && (f.dataset.type !== "race"));
-    context.features.forEach(f => {
+    const features = context.features.filter(f => (f.dataset.type !== "background") && (f.dataset.type !== "race"));
+    features.forEach(f => {
       if ( f.hasActions ) f.dataset.type = "active";
       else f.dataset.type = "passive";
     });
 
     // Add extra categories for features grouping.
     Object.values(this.actor.classes ?? {}).sort((a, b) => b.system.levels - a.system.levels).forEach(cls => {
-      context.features.push({
+      features.push({
         label: game.i18n.format("DND5E.FeaturesClass", { class: cls.name }),
         items: [],
         dataset: { type: cls.identifier }
@@ -521,126 +252,37 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
     });
 
     if ( this.actor.system.details.race instanceof dnd5e.documents.Item5e ) {
-      context.features.push({ label: "DND5E.FeaturesRace", items: [], dataset: { type: "race" } });
+      features.push({ label: "DND5E.FeaturesRace", items: [], dataset: { type: "race" } });
     }
 
     if ( this.actor.system.details.background instanceof dnd5e.documents.Item5e ) {
-      context.features.push({ label: "DND5E.FeaturesBackground", items: [], dataset: { type: "background" } });
+      features.push({ label: "DND5E.FeaturesBackground", items: [], dataset: { type: "background" } });
     }
 
-    context.features.push({ label: "DND5E.FeaturesOther", items: [], dataset: { type: "other" } });
-    context.classes = context.features.findSplice(f => f.isClass)?.items;
+    features.push({ label: "DND5E.FeaturesOther", items: [], dataset: { type: "other" } });
+    context.classes = features.findSplice(f => f.isClass)?.items;
 
-    // Spell slots
-    const plurals = new Intl.PluralRules(game.i18n.lang, { type: "ordinal" });
-    context.spellbook.forEach(section => {
-      if ( !section.usesSlots ) return;
-      const spells = foundry.utils.getProperty(this.actor.system.spells, section.prop);
-      const max = spells.override ?? spells.max ?? 0;
-      section.pips = Array.fromRange(max, 1).map(n => {
-        const filled = spells.value >= n;
-        const label = filled
-          ? game.i18n.format(`DND5E.SpellSlotN.${plurals.select(n)}`, { n })
-          : game.i18n.localize("DND5E.SpellSlotExpended");
-        const classes = ["pip"];
-        if ( filled ) classes.push("filled");
-        return { n, label, filled, tooltip: label, classes: classes.join(" ") };
-      });
+    context.features = {
+      sections: features,
+      filters: [
+        { key: "action", label: "DND5E.Action" },
+        { key: "bonus", label: "DND5E.BonusAction" },
+        { key: "reaction", label: "DND5E.Reaction" },
+        { key: "sr", label: "DND5E.ShortRest" },
+        { key: "lr", label: "DND5E.LongRest" },
+        { key: "concentration", label: "DND5E.Concentration" },
+        { key: "mgc", label: "DND5E.Item.Property.Magical" }
+      ]
+    };
+
+    // TODO: Customise this per-section.
+    features.forEach(section => {
+      section.categories = [
+        { classes: "item-uses", label: "DND5E.Uses", partial: "dnd5e.column-uses" },
+        { classes: "item-recovery", label: "DND5E.Recovery", partial: "dnd5e.column-recovery" },
+        { classes: "item-controls", partial: "dnd5e.column-feature-controls" }
+      ];
     });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @override */
-  _prepareItem(item, ctx) {
-    const { system } = item;
-
-    // Spells
-    if ( item.type === "spell" ) {
-
-      // Activation
-      const cost = system.activation?.cost;
-      const abbr = {
-        action: "DND5E.ActionAbbr",
-        bonus: "DND5E.BonusActionAbbr",
-        reaction: "DND5E.ReactionAbbr",
-        minute: "DND5E.TimeMinuteAbbr",
-        hour: "DND5E.TimeHourAbbr",
-        day: "DND5E.TimeDayAbbr"
-      }[system.activation.type];
-      ctx.activation = cost && abbr ? `${cost}${game.i18n.localize(abbr)}` : item.labels.activation;
-
-      // Range
-      const units = system.range?.units;
-      if ( units && (units !== "none") ) {
-        if ( units in CONFIG.DND5E.movementUnits ) {
-          ctx.range = {
-            distance: true,
-            value: system.range.value,
-            unit: game.i18n.localize(`DND5E.Dist${units.capitalize()}Abbr`)
-          };
-        }
-        else ctx.range = { distance: false };
-      }
-
-      // To Hit
-      const toHit = parseInt(item.labels.modifier);
-      if ( item.hasAttack && !isNaN(toHit) ) {
-        ctx.toHit = {
-          sign: Math.sign(toHit) < 0 ? "-" : "+",
-          abs: Math.abs(toHit)
-        };
-      }
-
-      // Prepared
-      const mode = system.preparation?.mode;
-      const config = CONFIG.DND5E.spellPreparationModes[mode] ?? {};
-      if ( config.prepares ) {
-        const isAlways = mode === "always";
-        const prepared = isAlways || system.preparation.prepared;
-        ctx.preparation = {
-          applicable: true,
-          disabled: !item.isOwner || isAlways,
-          cls: prepared ? "active" : "",
-          icon: `<i class="fa-${prepared ? "solid" : "regular"} fa-${isAlways ? "certificate" : "sun"}"></i>`,
-          title: isAlways
-            ? CONFIG.DND5E.spellPreparationModes.always.label
-            : prepared
-              ? CONFIG.DND5E.spellPreparationModes.prepared.label
-              : game.i18n.localize("DND5E.SpellUnprepared")
-        };
-      }
-      else ctx.preparation = { applicable: false };
-    }
-
-    // Gear
-    else {
-
-      // Attuned
-      if ( ctx.attunement ) {
-        ctx.attunement.applicable = true;
-        ctx.attunement.disabled = !item.isOwner;
-        ctx.attunement.cls = ctx.attunement.cls === "attuned" ? "active" : "";
-      }
-      else ctx.attunement = { applicable: false };
-
-      // Equipped
-      if ( "equipped" in system ) {
-        ctx.equip = {
-          applicable: true,
-          cls: system.equipped ? "active" : "",
-          title: `DND5E.${system.equipped ? "Equipped" : "Unequipped"}`,
-          disabled: !item.isOwner
-        };
-      }
-      else ctx.equip = { applicable: false };
-
-      // Subtitles
-      ctx.subtitle = [system.type?.label, item.isActive ? item.labels.activation : null].filterJoin(" &bull; ");
-    }
-
-    // Concentration
-    if ( this._concentration.items.has(item) ) ctx.concentration = true;
   }
 
   /* -------------------------------------------- */
@@ -648,15 +290,8 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   /** @inheritDoc */
   activateListeners(html) {
     super.activateListeners(html);
-    html.find(".pips[data-prop]").on("click", this._onTogglePip.bind(this));
     html.find(".death-tab").on("click", () => this._toggleDeathTray());
     html.find("[data-action]").on("click", this._onAction.bind(this));
-    html.find("[data-item-id][data-action]").on("click", this._onItemAction.bind(this));
-    html.find(".rollable:is(.saving-throw, .ability-check)").on("click", this._onRollAbility.bind(this));
-    html.find("proficiency-cycle").on("change", this._onChangeInput.bind(this));
-    html.find(".sidebar .collapser").on("click", this._onToggleSidebar.bind(this));
-    this.form.querySelectorAll(".item-tooltip").forEach(this._applyItemTooltips.bind(this));
-    this.form.querySelectorAll("[data-reference-tooltip]").forEach(this._applyReferenceTooltips.bind(this));
 
     // Prevent default middle-click scrolling when locking a tooltip.
     this.form.addEventListener("pointerdown", event => {
@@ -672,8 +307,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
     });
 
     if ( this.isEditable ) {
-      html.find(".meter > .hit-points").on("click", event => this._toggleEditHP(event, true));
-      html.find(".meter > .hit-points > input").on("blur", event => this._toggleEditHP(event, false));
       html.find(".create-child").on("click", this._onCreateChild.bind(this));
     }
 
@@ -717,10 +350,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   /** @inheritDoc */
   _onChangeTab(event, tabs, active) {
     super._onChangeTab(event, tabs, active);
-    this.form.className = this.form.className.replace(/tab-\w+/g, "");
-    this.form.classList.add(`tab-${active}`);
-    const sidebarCollapsed = game.user.getFlag("dnd5e", `sheetPrefs.character.tabs.${active}.collapseSidebar`);
-    if ( sidebarCollapsed !== undefined ) this._toggleSidebar(sidebarCollapsed);
     const createChild = this.form.querySelector(".create-child");
     createChild.setAttribute("aria-label", game.i18n.format("SIDEBAR.Create", {
       type: game.i18n.localize(`DOCUMENT.${active === "effects" ? "ActiveEffect" : "Item"}`)
@@ -769,59 +398,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   /* -------------------------------------------- */
 
   /**
-   * Handle the user toggling the sheet mode.
-   * @param {Event} event  The triggering event.
-   * @protected
-   */
-  async _onChangeSheetMode(event) {
-    const { MODES } = this.constructor;
-    const toggle = event.currentTarget;
-    const label = game.i18n.localize(`DND5E.SheetMode${toggle.checked ? "Play" : "Edit"}`);
-    toggle.dataset.tooltip = label;
-    toggle.setAttribute("aria-label", label);
-    this._mode = toggle.checked ? MODES.EDIT : MODES.PLAY;
-    await this.submit();
-    this.render();
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Handle toggling a pip on the character sheet.
-   * @param {PointerEvent} event  The triggering event.
-   * @returns {Promise<Actor5e>|void}
-   * @protected
-   */
-  _onTogglePip(event) {
-    const n = Number(event.target.closest("[data-n]")?.dataset.n);
-    if ( !n || isNaN(n) ) return;
-    const prop = event.currentTarget.dataset.prop;
-    let value = foundry.utils.getProperty(this.actor, prop);
-    if ( value === n ) value--;
-    else value = n;
-    return this.actor.update({ [prop]: value });
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Toggle editing hit points.
-   * @param {PointerEvent} event  The triggering event.
-   * @param {boolean} edit        Whether to toggle to the edit state.
-   * @protected
-   */
-  _toggleEditHP(event, edit) {
-    const target = event.currentTarget.closest(".hit-points");
-    const label = target.querySelector(":scope > .label");
-    const input = target.querySelector(":scope > input");
-    label.hidden = edit;
-    input.hidden = !edit;
-    if ( edit ) input.focus();
-  }
-
-  /* -------------------------------------------- */
-
-  /**
    * Toggle the death save tray.
    * @param {boolean} [open]  Force a particular open state.
    * @protected
@@ -833,38 +409,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
     this._deathTrayOpen = tray.classList.contains("open");
     tab.dataset.tooltip = `DND5E.DeathSave${this._deathTrayOpen ? "Hide" : "Show"}`;
     tab.setAttribute("aria-label", game.i18n.localize(tab.dataset.tooltip));
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Handle the user toggling the sidebar collapsed state.
-   * @protected
-   */
-  _onToggleSidebar() {
-    const collapsed = this._toggleSidebar();
-    const activeTab = this._tabs?.[0]?.active ?? "details";
-    game.user.setFlag("dnd5e", `sheetPrefs.character.tabs.${activeTab}.collapseSidebar`, collapsed);
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Toggle the sidebar collapsed state.
-   * @param {boolean} [collapsed]  Force a particular collapsed state.
-   * @returns {boolean}            The new collapsed state.
-   * @protected
-   */
-  _toggleSidebar(collapsed) {
-    this.form.classList.toggle("collapsed", collapsed);
-    collapsed = this.form.classList.contains("collapsed");
-    const collapser = this.form.querySelector(".sidebar .collapser");
-    const icon = collapser.querySelector("i");
-    collapser.dataset.tooltip = `JOURNAL.View${collapsed ? "Expand" : "Collapse"}`;
-    collapser.setAttribute("aria-label", game.i18n.localize(collapser.dataset.tooltip));
-    icon.classList.remove("fa-caret-left", "fa-caret-right");
-    icon.classList.add(`fa-caret-${collapsed ? "right" : "left"}`);
-    return collapsed;
   }
 
   /* -------------------------------------------- */
@@ -991,84 +535,6 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
   _onToggleSpellcasting(event) {
     const ability = event.currentTarget.closest("[data-ability]")?.dataset.ability;
     this.actor.update({ "system.attributes.spellcasting": ability });
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Initialize item tooltips on an element.
-   * @param {HTMLElement} element  The tooltipped element.
-   * @protected
-   */
-  _applyItemTooltips(element) {
-    if ( "tooltip" in element.dataset ) return;
-    const target = element.closest("[data-item-id], [data-effect-id], [data-uuid]");
-    let uuid = target.dataset.uuid;
-    if ( !uuid && target.dataset.itemId ) {
-      const item = this.actor.items.get(target.dataset.itemId);
-      uuid = item?.uuid;
-    } else if ( !uuid && target.dataset.effectId ) {
-      const { effectId, parentId } = target.dataset;
-      const collection = parentId ? this.actor.items.get(parentId).effects : this.actor.effects;
-      uuid = collection.get(effectId)?.uuid;
-    }
-    if ( !uuid ) return;
-    element.dataset.tooltip = `
-      <section class="loading" data-uuid="${uuid}"><i class="fas fa-spinner fa-spin-pulse"></i></section>
-    `;
-    element.dataset.tooltipClass = "dnd5e2 dnd5e-tooltip item-tooltip";
-    element.dataset.tooltipDirection ??= "LEFT";
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Initialize a rule tooltip on an element.
-   * @param {HTMLElement} element  The tooltipped element.
-   * @protected
-   */
-  _applyReferenceTooltips(element) {
-    if ( "tooltip" in element.dataset ) return;
-    const uuid = element.dataset.referenceTooltip;
-    element.dataset.tooltip = `
-      <section class="loading" data-uuid="${uuid}"><i class="fas fa-spinner fa-spin-pulse"></i></section>
-    `;
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Handle performing some action on an owned Item.
-   * @param {PointerEvent} event  The triggering event.
-   * @protected
-   */
-  _onItemAction(event) {
-    if ( event.target.closest("select") ) return;
-    event.preventDefault();
-    event.stopPropagation();
-    const itemId = event.currentTarget.closest("[data-item-id]")?.dataset.itemId;
-    const action = event.currentTarget.dataset.action;
-    const item = this.actor.items.get(itemId);
-
-    switch ( action ) {
-      case "edit": item?.sheet.render(true); break;
-      case "delete": item?.deleteDialog(); break;
-    }
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Handle rolling an ability check or saving throw.
-   * @param {PointerEvent} event  The triggering event.
-   * @protected
-   */
-  _onRollAbility(event) {
-    const abilityId = event.currentTarget.closest("[data-ability]").dataset.ability;
-    const isSavingThrow = event.currentTarget.classList.contains("saving-throw");
-    if ( abilityId === "concentration" ) this.actor.rollConcentration({ event });
-    else if ( isSavingThrow ) this.actor.rollAbilitySave(abilityId, { event });
-    else this.actor.rollAbilityTest(abilityId, { event });
   }
 
   /* -------------------------------------------- */
@@ -1260,7 +726,7 @@ export default class ActorSheet5eCharacter2 extends ActorSheet5eCharacter {
       if ( uses?.max > 100 ) css.push("uses-sm");
       if ( modifier !== undefined ) {
         const value = Number(modifier.replace?.(/\s+/g, "") ?? modifier);
-        if ( !isNaN(value) ) modifier = { abs: Math.abs(value), sign: value < 0 ? "-" : "+" };
+        if ( !isNaN(value) ) modifier = value;
       }
 
       const rollableClass = [];
