+++ templates/actors/parts/columns/column-formula.hbs
@@ -1,0 +1,12 @@
+<div class="item-detail item-formula {{#unless item.labels.derivedDamage}}empty{{/unless}}">
+    {{#each item.labels.derivedDamage}}
+    <div class="row">
+        <span class="formula">{{ formula }}</span>
+        {{#with (lookup @root.labels.damageAndHealing damageType)}}
+        <span data-tooltip="{{ label }}" aria-label="{{ label }}">
+            <dnd5e-icon src="{{ icon }}"></dnd5e-icon>
+        </span>
+        {{/with}}
+    </div>
+    {{/each}}
+</div>
