+++ templates/actors/parts/columns/column-recovery.hbs
@@ -1,0 +1,5 @@
+<div class="item-detail item-recovery {{#unless item.system.uses.per}}empty{{/unless}}">
+    {{#if item.system.uses.per}}
+    <span class="condensed">{{ item.labels.recovery }}</span>
+    {{/if}}
+</div>
