+++ templates/chat/item-card.hbs
@@ -145,12 +145,12 @@
     {{#if data.hasProperties}}
     <ul class="card-footer pills unlist">
         {{#each data.tags}}
-        <li class="pill pill-sm">
+        <li class="pill transparent">
             <span class="label">{{ this }}</span>
         </li>
         {{/each}}
         {{#each data.properties}}
-        <li class="pill pill-sm">
+        <li class="pill transparent">
             <span class="label">{{ this }}</span>
         </li>
         {{/each}}
