+++ templates/effects/parts/effect-tooltip.hbs
@@ -20,7 +20,7 @@
     <section class="description">{{{ description }}}</section>
     <ul class="pills">
         {{#each properties}}
-        <li class="pill pill-sm">
+        <li class="pill transparent">
             <span class="label">{{ this }}</span>
         </li>
         {{/each}}
