+++ templates/items/parts/item-summary.hbs
@@ -1,7 +1,7 @@
 <div class="item-summary">
     {{{description.value}}}
-  
-    <div class="item-properties">
-        {{#each properties}}<span class="tag">{{this}}</span>{{/each}}
+
+    <div class="item-properties pills">
+        {{#each properties}}<span class="tag pill transparent pill-xs">{{this}}</span>{{/each}}
     </div>
 </div>
