+++ templates/items/parts/item-tooltip.hbs
@@ -76,13 +76,13 @@
     <ul class="pills">
         {{#if isSpell}}
         {{#each tags}}
-        <li class="pill pill-sm">
+        <li class="pill transparent">
             <span class="label">{{ this }}</span>
         </li>
         {{/each}}
         {{else}}
         {{#each properties}}
-        <li class="pill pill-sm">
+        <li class="pill transparent">
             <span class="label">{{ this }}</span>
         </li>
         {{/each}}
